github desktop learning
not sure what is going on with github desktop. I am cloning from appbrewery files to github desktop, working xcode and this window looks like it's still open and asking me to commit to master. I thought saving from xcode in my local folder is enough for my purpose. Should I not be saving to github desktop if I am not looking to contribute to anothers project?